Propagate refresh token error #259
https://github.com/plamere/spotipy/issues/259
