+++ spotipy/__init__.py
@@ -1,2 +1,3 @@
 VERSION='2.0.1'
-from .client import Spotify, SpotifyException
+from .client import Spotify
+from .exceptions import SpotifyException
